--- src/components/Omniture.js
@@ -1,10 +1,0 @@
-import useScript from 'hooks/useScript';
-
-const Omniture = () => {
-
-    useScript('//image.sony.co.kr/omniture/real/sonystore_code_2013.js');
-
-    return <></>;
-};
-
-export default Omniture;
