+++ src/components/SEOHelmet.js
@@ -1,17 +1,17 @@
 import { Helmet } from 'react-helmet';
 
+import useScript from 'hooks/useScript';
+
 const SEOHelmet = ({ title, children }) => {
+    
+
     return (
-        <Helmet>
-            <title>{title}</title>
-            {children}
-            {title && (
-                <script
-                    defer
-                    src='//image.sony.co.kr/omniture/omni_dev/sonystore_code_2013.js'
-                ></script>
-            )}
-        </Helmet>
+        <>
+            <Helmet>
+                <title>{title}</title>
+                {children}
+            </Helmet>
+        </>
     );
 };
