+++ src/components/common/Tabs.js
@@ -1,5 +1,4 @@
 import { useEffect } from 'react';
-import {OnclickTracker} from "../ReactGA4Tracker";
 
 export default function Tabs({ tabs, tabState, setTabState, defaultState }) {
   useEffect(() => {
@@ -16,7 +15,6 @@ export default function Tabs({ tabs, tabState, setTabState, defaultState }) {
             onClick={(event) => {
               event.preventDefault();
               setTabState(tabName);
-              OnclickTracker(tabState, tabState);
             }}
           >
             {label}
