+++ src/components/event/SectionMap.js
@@ -117,9 +117,7 @@ const SectionMap = ({ product, gift, giftProduct, addToCart, hsValidator }) => {
                         onClick={() => addToCart(product.productNo)}
                         disabled={isSoldOut}
                     >
-
                         {isSoldOut ? '일시품절' : win < 640 ? '장바구니' : onlyMo ? '장바구니' : '장바구니 담기'}
-
                     </button>
                     <HsValidator ref={hsValidator} />
                 </div>
